+++ samples/hello/hello_combo.rb
@@ -46,7 +46,7 @@ class HelloCombo
       text 'Hello, Combo!'
       
       combo(:read_only) {
-        selection bind(person, :country)
+        selection bind(person, :country) # also binds to country_options by convention
       }
       
       button {